requirements_frontend.txt
streamlit
torch
torchvision
opencv-contrib-python
numpy
pillow
matplotlib
tqdm
